Remove Includes and Fix Errors
Replace Make_book, Update README
