+++ README
@@ -28,12 +28,9 @@ regexp.html         The appendix on regular expressions
 browser.html        The appendix on web browsers
 css.html            The appendix on Cascading Style Sheets
 cyc.js              A simple Cyc processor
-include.js          A simple @include preprocessor
-include.c           A simple @include preprocessor in C
 onehtml.js          Rules for outputing a single HTML file
 make_book.js        Make a book as a single HTML file
 
-It also makes use of RQ: https://github.com/douglascrockford/RQ
 
 You can make a single file encyclopedia with
