+++ x.cyc
@@ -2,7 +2,7 @@
 
 @specimen "XML"
 
-    XML @aka 'Extensible Markup Language' is a document format based on SGML.
+    XML @aka(Extensible Markup Language) is a document format based on SGML.
     It is often misused as a data format. See @url<http://www.w3.org/TR/REC-xml>.
 
 @specimen "xor @aka(exclusive or)"
